requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
ccxt==4.3.54
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
kombu==5.3.7
multidict==6.0.5
prompt_toolkit==3.0.47
pycares==4.4.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
